requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
Pillow==11.0.0
plotly==5.9.0
seaborn==0.13.2
streamlit==1.26.0
